backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiomysql==0.2.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-dotenv==1.0.0

# AI & ML
google-generativeai==0.3.2
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80

# HTTP Client & Caching
httpx==0.26.0
redis==5.0.1
certifi==2024.2.2
python-certifi-win32==1.6.1  # Merges Windows Certificate Store with certifi

# Rate Limiting
slowapi==0.1.9

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
aiofiles==23.2.1
